sandbox - share most desktop.main code
